Move fake_owner to control module and clean up.
